+++ DudeClothing/src/main/java/com/dudeclothing/admin/model/Client.java
@@ -14,20 +14,18 @@ public class Client {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int clientId;
-
+	private long clientId;
 	private String name;
-
 	private int phoneNo;
 
 	public Client() {
 	}
 
-	public int getClientId() {
+	public long getClientId() {
 		return this.clientId;
 	}
 
-	public void setClientId(int clientId) {
+	public void setClientId(long clientId) {
 		this.clientId = clientId;
 	}
