+++ DudeClothing/src/main/java/com/dudeclothing/admin/model/Status.java
@@ -14,18 +14,17 @@ public class Status {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int statusId;
-
+	private long statusId;
 	private String name;
 
 	public Status() {
 	}
 
-	public int getStatusId() {
+	public long getStatusId() {
 		return this.statusId;
 	}
 
-	public void setStatusId(int statusId) {
+	public void setStatusId(long statusId) {
 		this.statusId = statusId;
 	}
